@@ -665,7 +665,7 @@ if (isset($_GET["mongo"])) {
 	function convert_field($field) {
 	}
 
-	function unconvert_field($field, $return) {
+	function unconvert_field(array $field, $return) {
 		return $return;
 	}
 
