@@ -1148,7 +1148,7 @@ if (isset($_GET["mysql"])) {
 	* @param string
 	* @return string
 	*/
-	function unconvert_field($field, $return) {
+	function unconvert_field(array $field, $return) {
 		if (preg_match("~binary~", $field["type"])) {
 			$return = "UNHEX($return)";
 		}
@@ -1159,6 +1159,7 @@ if (isset($_GET["mysql"])) {
 			$prefix = (min_version(8) ? "ST_" : "");
 			$return = $prefix . "GeomFromText($return, $prefix" . "SRID($field[field]))";
 		}
+
 		return $return;
 	}
 
