@@ -93,7 +93,7 @@ $j = 0;
 foreach ($row["source"] as $key => $val) {
 	echo "<tr>";
 	echo "<td>" . html_select("source[" . (+$key) . "]", array(-1 => "") + $source, $val, ($j == count($row["source"]) - 1 ? "foreignAddRow.call(this);" : 1), "label-source");
-	echo "<td>" . html_select("target[" . (+$key) . "]", $target, $row["target"][$key], 1, "label-target");
+	echo "<td>" . html_select("target[" . (+$key) . "]", $target, $row["target"][$key] ?? null, 1, "label-target");
 	$j++;
 }
 ?>
