@@ -111,7 +111,7 @@ function check_invalid_login() {
 	$invalids = file_exists($filename) ? unserialize(file_get_contents($filename)) : [];
 	$invalid = ($invalids ? $invalids[$adminer->bruteForceKey()] : []);
 
-	$next_attempt = ($invalid[1] > 29 ? $invalid[0] - time() : 0); // allow 30 invalid attempts
+	$next_attempt = ($invalid && $invalid[1] > 29 ? $invalid[0] - time() : 0); // allow 30 invalid attempts
 	if ($next_attempt > 0) { //! do the same with permanent login
 		auth_error(lang('Too many unsuccessful logins, try again in %d minute(s).', ceil($next_attempt / 60)));
 	}
