@@ -183,7 +183,7 @@ function get_nonce()
 */
 function page_messages($error) {
 	$uri = preg_replace('~^[^?]*~', '', $_SERVER["REQUEST_URI"]);
-	$messages = $_SESSION["messages"][$uri];
+	$messages = $_SESSION["messages"][$uri] ?? null;
 	if ($messages) {
 		echo "<div class='message'>" . implode("</div>\n<div class='message'>", $messages) . "</div>" . script("messagesPrint();");
 		unset($_SESSION["messages"][$uri]);
