@@ -25,12 +25,12 @@ function xhtml_open_tags($s) {
 $coverage_filename = sys_get_temp_dir() . "/adminer_coverage.ser";
 if (!extension_loaded("xdebug")) {
 	echo "<p class='error'>Xdebug has to be enabled.</p>\n";
-} elseif ($_GET["coverage"] === "0") {
+} elseif (($_GET["coverage"] ?? null) === "0") {
 	file_put_contents($coverage_filename, serialize(array()));
 	echo "<p class='message'>Coverage started.</p>\n";
-} elseif (preg_match('~^(adminer|editor)/(include/)?[-_.a-z0-9]+$~i', $_GET["coverage"])) {
+} elseif (preg_match('~^(adminer|editor)/(include/)?[-_.a-z0-9]+$~i', $_GET["coverage"] ?? null)) {
 	// highlight single file
-	$filename = $_GET["coverage"];
+	$filename = $_GET["coverage"] ?? null;
 	$coverage = (file_exists($coverage_filename) ? unserialize(file_get_contents($coverage_filename)) : array());
 	$file = explode("<br />", highlight_file($filename, true));
 	$prev_color = null;
