@@ -397,7 +397,7 @@ if (isset($_GET["simpledb"])) {
 		return " $query$where" . ($limit !== null ? $separator . "LIMIT $limit" : "");
 	}
 
-	function unconvert_field($field, $return) {
+	function unconvert_field(array $field, $return) {
 		return $return;
 	}
 
