@@ -21,13 +21,13 @@ HTML;
 		return null;
 	}
 
-	function processInput($field, $value, $function = '') {
+	function processInput(?array $field, $value, $function = '') {
 		if ($function === '') {
 			$json = $this->_testJson($value);
 			if ($json !== $value) {
 				$value = json_encode($json);
 			}
 		}
-		return adminer()->_callParent('processInput', array($field, $value, $function));
+		return adminer()->_callParent('processInput', [$field, $value, $function]);
 	}
 }
